mc_pos_control: split out header and name consistently
