+++ src/modules/mc_pos_control/CMakeLists.txt
@@ -1,6 +1,6 @@
 ############################################################################
 #
-#   Copyright (c) 2015 PX4 Development Team. All rights reserved.
+#   Copyright (c) 2015-2020 PX4 Development Team. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions
@@ -40,7 +40,8 @@ px4_add_module(
 	COMPILE_FLAGS
 		-Wno-implicit-fallthrough # TODO: fix and remove
 	SRCS
-		mc_pos_control_main.cpp
+		MulticopterPositionControl.cpp
+		MulticopterPositionControl.hpp
 	DEPENDS
 		PositionControl
 		controllib
